V2 integration

+++ Warehousing/DataControllers_v2/ItemGroupController.cs
@@ -5,7 +5,7 @@ using Warehousing.DataServices_v2;
 namespace Warehousing.DataControllers_v2
 {
 [ApiController]
-[Route("api/itemgroups/v2")]
+[Route("api/v2/itemgroups")]
 public class ItemGroupsController : ControllerBase
 {
     private readonly ItemGroupService _itemGroupService;
@@ -27,8 +27,14 @@ public class ItemGroupsController : ControllerBase
     }
 
     [HttpPost]
-    public IActionResult AddItemGroup([FromBody] ItemGroup itemGroup)
+    public IActionResult Post([FromBody] ItemGroup itemGroup)
     {
+        var itemGroups = _itemGroupService.GetAllItemGroups();
+        if (itemGroups.Any(ig => ig.Name == itemGroup.Name))
+        {
+            return BadRequest($"ItemGroup with Name {itemGroup.Name} already exists");
+        }
+
         _itemGroupService.AddItemGroup(itemGroup);
         return CreatedAtAction(nameof(GetItemGroupById), new { id = itemGroup.Id }, itemGroup);
     }
@@ -36,15 +42,28 @@ public class ItemGroupsController : ControllerBase
     [HttpPut("{id}")]
     public IActionResult UpdateItemGroup(int id, [FromBody] ItemGroup updatedItemGroup)
     {
+        var itemGroup = _itemGroupService.GetItemGroupById(id);
+        if (itemGroup == null) return NotFound();
+
+        var itemGroups = _itemGroupService.GetAllItemGroups();
+
+        if (itemGroups.Any(ig => ig.Name == updatedItemGroup.Name && ig.Id != id))
+        {
+            return BadRequest($"ItemGroup with Name {itemGroup.Name} already exists");
+        }
+
         _itemGroupService.UpdateItemGroup(id, updatedItemGroup);
-        return NoContent();
+        return Ok(updatedItemGroup);
     }
 
     [HttpDelete("{id}")]
     public IActionResult DeleteItemGroup(int id)
     {
+        var itemGroup = _itemGroupService.GetItemGroupById(id);
+        if (itemGroup == null) return NotFound();
+
         _itemGroupService.DeleteItemGroup(id);
-        return NoContent();
+        return Ok();
     }
 }
 }
